remove comments & fix disabled hotkeys

+++ src/components/DeleteTodo.vue
@@ -17,7 +17,7 @@ v-dialog(v-model='dialog', persistent, scrollable, max-width='600px')
         @click='deleteTodo',
         :loading='loading',
         :disabled='loading',
-        v-hotkey='keymap',
+        v-hotkey='keymap'
       ) {{ $t("delete") }}
 </template>
 
@@ -35,6 +35,7 @@ import { User } from '@/models/User'
 
 const UserStore = namespace('UserStore')
 const SnackbarStore = namespace('SnackbarStore')
+const SettingsStore = namespace('SettingsStore')
 
 @Component
 export default class DeleteTodo extends Vue {
@@ -42,13 +43,14 @@ export default class DeleteTodo extends Vue {
 
   @UserStore.State user?: User
   @SnackbarStore.Mutation setSnackbarError!: (error: string) => void
+  @SettingsStore.State hotKeysEnabled!: boolean
 
   loading = false
   dialog = false
 
   get keymap() {
     return {
-      'enter': this.deleteTodo
+      enter: () => this.deleteTodo(true),
     }
   }
 
@@ -57,7 +59,8 @@ export default class DeleteTodo extends Vue {
     this.dialog = !!val
   }
 
-  async deleteTodo() {
+  async deleteTodo(hotkey = false) {
+    if (hotkey && !this.hotKeysEnabled) return
     this.loading = true
     try {
       await api.deleteTodo((this as any).todo)
